Don't subclass the chatbox for the controlbox model
